refactor: 안쓰는 코드 삭제 및 user 패키지 변경
