+++ src/main/java/com/softeer/podo/admin/repository/ArrivalUserRepository.java
@@ -1,6 +1,6 @@
-package com.softeer.podo.user.repository;
+package com.softeer.podo.admin.repository;
 
-import com.softeer.podo.user.model.entity.ArrivalUser;
+import com.softeer.podo.admin.model.entity.ArrivalUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArrivalUserRepository extends JpaRepository<ArrivalUser, Long> {
