+++ src/main/java/com/softeer/podo/admin/service/AdminService.java
@@ -13,9 +13,9 @@ import com.softeer.podo.admin.model.exception.EventNotFoundException;
 import com.softeer.podo.admin.repository.EventRepository;
 import com.softeer.podo.admin.repository.EventRewardRepository;
 import com.softeer.podo.admin.repository.EventWeightRepository;
-import com.softeer.podo.user.model.entity.LotsUser;
-import com.softeer.podo.user.repository.ArrivalUserRepository;
-import com.softeer.podo.user.repository.LotsUserRepository;
+import com.softeer.podo.admin.model.entity.LotsUser;
+import com.softeer.podo.admin.repository.ArrivalUserRepository;
+import com.softeer.podo.admin.repository.LotsUserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
