+++ src/main/java/com/softeer/podo/event/service/EventLotsService.java
@@ -16,8 +16,8 @@ import com.softeer.podo.event.repository.TestResultRepository;
 import com.softeer.podo.event.util.Result;
 import com.softeer.podo.event.util.SelectionMap;
 import com.softeer.podo.security.AuthInfo;
-import com.softeer.podo.user.model.entity.LotsUser;
-import com.softeer.podo.user.repository.LotsUserRepository;
+import com.softeer.podo.admin.model.entity.LotsUser;
+import com.softeer.podo.admin.repository.LotsUserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
