+++ src/main/java/com/softeer/podo/security/jwt/TokenProvider.java
@@ -7,7 +7,7 @@ import com.nimbusds.jose.crypto.MACSigner;
 import com.nimbusds.jose.crypto.MACVerifier;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
-import com.softeer.podo.user.model.entity.Role;
+import com.softeer.podo.admin.model.entity.Role;
 import com.softeer.podo.security.jwt.exception.InvalidTokenException;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
