+++ src/main/java/com/softeer/podo/test/repository/TestUserRepository.java
@@ -1,6 +1,6 @@
 package com.softeer.podo.test.repository;
 
-import com.softeer.podo.user.model.entity.TestUser;
+import com.softeer.podo.admin.model.entity.TestUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TestUserRepository extends JpaRepository<TestUser, Long> {
