--- src/main/java/com/softeer/podo/user/model/dto/UserDto.java
@@ -1,18 +1,0 @@
-package com.softeer.podo.user.model.dto;
-
-import com.softeer.podo.user.model.entity.Role;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserDto {
-    private Long id;
-    private String name;
-    private String phoneNum;
-    private Role role;
-}
